requirements.txt
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
google-api-core==2.23.0
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-translate==3.17.0
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.13.1
grpcio==1.67.1
grpcio-status==1.67.1
httplib2==0.22.0
idna==3.10
linkify-it-py==2.0.3
markdown-it-py==3.0.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
oauthlib==3.2.2
platformdirs==4.3.6
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.18.0
pynput==1.7.7
pyobjc-core==10.3.1
pyobjc-framework-ApplicationServices==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-CoreText==10.3.1
pyobjc-framework-Quartz==10.3.1
pyparsing==3.2.0
readchar==4.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
six==1.16.0
textual==0.86.3
typing_extensions==4.12.2
uc-micro-py==1.0.3
uritemplate==4.1.1
urllib3==2.2.3
